requirements.txt
py4j==0.10.9.5
numpy   
pandas
Django
requests
sqlalchemy<2.0
finance-datareader
quandl
billiard
yfinance
boto3
psycopg2-binary
python-dotenv
redshift_connector
billiard